refactor: combined React logic into the Feedback components - removed Multiline files
