+++ streamlit_feedback/frontend/src/FacesWithQualiFeedback.js
@@ -5,10 +5,12 @@ import SentimentNeutralIcon from '@mui/icons-material/SentimentNeutral';
 import SentimentSatisfiedIcon from '@mui/icons-material/SentimentSatisfied';
 import SentimentSatisfiedAltIcon from '@mui/icons-material/SentimentSatisfiedAlt';
 import InputBase from '@mui/material/InputBase';
-import { styled } from '@mui/system';
 import Button from '@mui/material/Button';
 import Stack from '@mui/material/Stack';
 import { Box } from "@mui/material";
+
+import { styled } from '@mui/material/styles';
+import TextField from "@mui/material/TextField";
 
 const colors = {
     grey: "#c7d1d3",
@@ -19,8 +21,16 @@ const colors = {
     "😞": "#f44336"
 }
 
+const TextFieldcolors = {
+    "😀": "success",
+    "🙂": "success",
+    "😐": "warning",
+    "🙁": "error",
+    "😞": "error"
+}
+
 const StyledCustomInput = styled(InputBase)(
-({ color }) => `
+    ({ color }) => `
     width: 60vw;
     font-family: sans-serif;
     font-size: 0.875rem;
@@ -33,13 +43,27 @@ const StyledCustomInput = styled(InputBase)(
     `
 );
 
+const StyledTextField = styled(TextField)(
+    ({ color }) => `
+        width: 60vw;
+        font-family: sans-serif;
+        font-size: 0.875rem;
+        font-weight: 400;
+        padding: 0px 12px;
+        border-radius: 8px;
+        color: ${color};
+        border: 1px solid ${color};
+        background: transparent;
+        `
+);
+
 export function FacesWithQualiFeedback(props) {
     const [submitted, setSubmitted] = useState(false);
     const [inputText, setInputText] = useState(null);
     const [faceScore, setFaceScore] = useState(null);
 
     useEffect(() => {
-        if (props.disableWithScore){
+        if (props.disableWithScore) {
             setSubmitted(true);
             setFaceScore(props.disableWithScore);
         }
@@ -68,7 +92,7 @@ export function FacesWithQualiFeedback(props) {
             return colors["grey"]
         }
     }
-    
+
     const selectHoverColor = (score) => {
         if (faceScore) {
             if (score === faceScore) {
@@ -94,62 +118,136 @@ export function FacesWithQualiFeedback(props) {
         props.submitFeedback(faceScore, inputText);
     };
 
-    return (
-        <Box paddingY={0.5}>
-            <Stack direction="row" spacing={1} justifyContent={props.align}>
-                <SentimentVeryDissatisfiedIcon
-                sx={{
-                    fontSize: 28,
-                    color: selectColor("😞"),
-                    '&:hover': {
-                        cursor: submitted ? null : "pointer",
-                        color: selectHoverColor("😞"),
-                    }, }}
-                onClick={() => submitted ? {} : handleFaceClick("😞")}
-                />
-                <SentimentDissatisfiedIcon
-                sx={{
-                    fontSize: 28,
-                    color: selectColor("🙁"),
-                    '&:hover': {
-                        cursor: submitted ? null : "pointer",
-                        color: selectHoverColor("🙁"),
-                    }, }}
-                onClick={() => submitted ? {} : handleFaceClick("🙁")}
-                />
-                <SentimentNeutralIcon
-                sx={{
-                    fontSize: 28,
-                    color: selectColor("😐"),
-                    '&:hover': {
-                        cursor: submitted ? null : "pointer",
-                        color: selectHoverColor("😐"),
-                    }, }}
-                onClick={() => submitted ? {} : handleFaceClick("😐")}
-                />
-                <SentimentSatisfiedIcon
-                sx={{
-                    fontSize: 28,
-                    color: selectColor("🙂"),
-                    '&:hover': {
-                        cursor: submitted ? null : "pointer",
-                        color: selectHoverColor("🙂"),
-                    }, }}
-                onClick={() => submitted ? {} : handleFaceClick("🙂")}
-                />
-                <SentimentSatisfiedAltIcon
-                sx={{
-                    fontSize: 28,
-                    color: selectColor("😀"),
-                    '&:hover': {
-                        cursor: submitted ? null : "pointer",
-                        color: selectHoverColor("😀"),
-                    }, }}
-                onClick={() => submitted ? {} : handleFaceClick("😀")}
-                />
-                {submitted === false && faceScore !== null ? <StyledCustomInput onChange={handleTextInput} aria-label="Demo input" placeholder={props.optionalTextLabel} color={colors[faceScore]}/> : null}
-                {submitted === false && faceScore !== null ? <Button sx={{color: colors[faceScore]}} variant="text" size="small" onClick={handleSubmission}>Submit</Button> : null}
-            </Stack>
-        </Box>
+
+    if (props.maxTextLength != null) {
+        return (
+            <Box paddingY={0.5} height={140} component="form" sx={{ "& .MuiTextField-root": { m: 1, width: "50ch" } }} noValidate autoComplete="off">
+                <Stack direction="row" spacing={1} justifyContent={props.align}>
+                    <SentimentVeryDissatisfiedIcon
+                        sx={{
+                            fontSize: 28,
+                            color: selectColor("😞"),
+                            '&:hover': {
+                                cursor: submitted ? null : "pointer",
+                                color: selectHoverColor("😞"),
+                            },
+                        }}
+                        onClick={() => submitted ? {} : handleFaceClick("😞")}
+                    />
+                    <SentimentDissatisfiedIcon
+                        sx={{
+                            fontSize: 28,
+                            color: selectColor("🙁"),
+                            '&:hover': {
+                                cursor: submitted ? null : "pointer",
+                                color: selectHoverColor("🙁"),
+                            },
+                        }}
+                        onClick={() => submitted ? {} : handleFaceClick("🙁")}
+                    />
+                    <SentimentNeutralIcon
+                        sx={{
+                            fontSize: 28,
+                            color: selectColor("😐"),
+                            '&:hover': {
+                                cursor: submitted ? null : "pointer",
+                                color: selectHoverColor("😐"),
+                            },
+                        }}
+                        onClick={() => submitted ? {} : handleFaceClick("😐")}
+                    />
+                    <SentimentSatisfiedIcon
+                        sx={{
+                            fontSize: 28,
+                            color: selectColor("🙂"),
+                            '&:hover': {
+                                cursor: submitted ? null : "pointer",
+                                color: selectHoverColor("🙂"),
+                            },
+                        }}
+                        onClick={() => submitted ? {} : handleFaceClick("🙂")}
+                    />
+                    <SentimentSatisfiedAltIcon
+                        sx={{
+                            fontSize: 28,
+                            color: selectColor("😀"),
+                            '&:hover': {
+                                cursor: submitted ? null : "pointer",
+                                color: selectHoverColor("😀"),
+                            },
+                        }}
+                        onClick={() => submitted ? {} : handleFaceClick("😀")}
+                    />
+                    {submitted === false && faceScore !== null ? <StyledTextField id="outlined-multiline-static" inputProps={{ maxLength: props.maxTextLength }} onChange={handleTextInput} multiline rows={4} placeholder={props.optionalTextLabel} aria-label="Demo input" color={TextFieldcolors[faceScore]} /> : null}
+                    {submitted === false && faceScore !== null ? <Button sx={{ color: colors[faceScore] }} variant="text" size="small" onClick={handleSubmission}>Submit</Button> : null}
+                </Stack>
+            </Box>
         )
     }
+    else {
+
+        return (
+            <Box paddingY={0.5}>
+                <Stack direction="row" spacing={1} justifyContent={props.align}>
+                    <SentimentVeryDissatisfiedIcon
+                        sx={{
+                            fontSize: 28,
+                            color: selectColor("😞"),
+                            '&:hover': {
+                                cursor: submitted ? null : "pointer",
+                                color: selectHoverColor("😞"),
+                            },
+                        }}
+                        onClick={() => submitted ? {} : handleFaceClick("😞")}
+                    />
+                    <SentimentDissatisfiedIcon
+                        sx={{
+                            fontSize: 28,
+                            color: selectColor("🙁"),
+                            '&:hover': {
+                                cursor: submitted ? null : "pointer",
+                                color: selectHoverColor("🙁"),
+                            },
+                        }}
+                        onClick={() => submitted ? {} : handleFaceClick("🙁")}
+                    />
+                    <SentimentNeutralIcon
+                        sx={{
+                            fontSize: 28,
+                            color: selectColor("😐"),
+                            '&:hover': {
+                                cursor: submitted ? null : "pointer",
+                                color: selectHoverColor("😐"),
+                            },
+                        }}
+                        onClick={() => submitted ? {} : handleFaceClick("😐")}
+                    />
+                    <SentimentSatisfiedIcon
+                        sx={{
+                            fontSize: 28,
+                            color: selectColor("🙂"),
+                            '&:hover': {
+                                cursor: submitted ? null : "pointer",
+                                color: selectHoverColor("🙂"),
+                            },
+                        }}
+                        onClick={() => submitted ? {} : handleFaceClick("🙂")}
+                    />
+                    <SentimentSatisfiedAltIcon
+                        sx={{
+                            fontSize: 28,
+                            color: selectColor("😀"),
+                            '&:hover': {
+                                cursor: submitted ? null : "pointer",
+                                color: selectHoverColor("😀"),
+                            },
+                        }}
+                        onClick={() => submitted ? {} : handleFaceClick("😀")}
+                    />
+                    {submitted === false && faceScore !== null ? <StyledCustomInput onChange={handleTextInput} aria-label="Demo input" placeholder={props.optionalTextLabel} color={colors[faceScore]} /> : null}
+                    {submitted === false && faceScore !== null ? <Button sx={{ color: colors[faceScore] }} variant="text" size="small" onClick={handleSubmission}>Submit</Button> : null}
+                </Stack>
+            </Box>
+        )
+    }
+}
